+++ src/inspect_ai/_view/www/src/client/api/client-api.ts
@@ -44,6 +44,7 @@ export class SampleSizeLimitedExceededError extends Error {
     Object.setPrototypeOf(this, SampleSizeLimitedExceededError.prototype);
   }
 }
+
 interface LoadedLogFile {
   file?: string;
   remoteLog?: RemoteLogFile;
