+++ src/inspect_ai/_view/www/src/client/api/index.ts
@@ -1,11 +1,11 @@
 import JSON5 from "json5";
 import { dirname } from "../../utils/path";
 import { getVscodeApi } from "../../utils/vscode";
-import { createViewServerApi } from "./api-view-server";
-import simpleHttpApi from "./api-http";
-import vscodeApi from "./api-vscode";
 import { clientApi } from "./client-api";
+import staticHttpApi from "./static-http/api-static-http";
 import { ClientAPI } from "./types";
+import { viewServerApi } from "./view-server/api-view-server";
+import vscodeApi from "./vscode/api-vscode";
 
 /**
  * Resolves the client API
@@ -25,7 +25,7 @@ const resolveApi = (): ClientAPI => {
         const data = JSON5.parse(context);
         if (data.log_dir || data.log_file) {
           const log_dir = data.log_dir || dirname(data.log_file);
-          const api = simpleHttpApi(log_dir, data.log_file);
+          const api = staticHttpApi(log_dir, data.log_file);
           return clientApi(api, data.log_file);
         }
       }
@@ -42,21 +42,21 @@ const resolveApi = (): ClientAPI => {
 
     if (forceViewServerApi) {
       return clientApi(
-        createViewServerApi({ logDir: resolved_log_dir }),
+        viewServerApi({ logDir: resolved_log_dir }),
         resolved_log_file,
       );
     }
 
     if (resolved_log_dir !== undefined || resolved_log_file !== undefined) {
       return clientApi(
-        simpleHttpApi(resolved_log_dir, resolved_log_file),
+        staticHttpApi(resolved_log_dir, resolved_log_file),
         resolved_log_file,
       );
     }
 
     // No signal information so use the standard
     // view server API (inspect view)
-    return clientApi(createViewServerApi());
+    return clientApi(viewServerApi());
   }
 };
