+++ src/inspect_ai/_view/www/src/index.ts
@@ -3,11 +3,11 @@ export { App } from "./app/App";
 
 // Client APIs
 export { clientApi } from "./client/api/client-api";
-export { createViewServerApi } from "./client/api/api-view-server.ts";
-export { default as simpleHttpApi } from "./client/api/api-http";
+export { default as simpleHttpApi } from "./client/api/static-http/api-static-http.ts";
+export { viewServerApi as createViewServerApi } from "./client/api/view-server/api-view-server.ts";
 
 // Client API - Types
-export type { ClientAPI, LogViewAPI, Capabilities } from "./client/api/types";
+export type { Capabilities, ClientAPI, LogViewAPI } from "./client/api/types";
 
 // State Store
 export { initializeStore } from "./state/store";
